# lib/api-zod/src/generated/types/updateNoteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Notes API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateNoteInput {
  title?: string;
  content?: string;
}